Updated offline sim script
